Web/requirements.txt
bidict==0.22.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dlib==20.0.0
dnspython==2.4.2
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
motor==3.3.2
numpy==1.26.2
Pillow==10.1.0
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.0
python-engineio==4.8.1
python-socketio==5.10.0
simple-websocket==1.0.0
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.1
